RuntimeSession接口新增测试状态更新事件。修改执行时长数据类型为double

+++ source/src/Dev/Common/Runtime/ISequenceTestResult.cs
@@ -44,7 +44,7 @@ namespace Testflow.Runtime
         /// <summary>
         /// 消耗的时间，单位为ms
         /// </summary>
-        ulong ElapsedTime { get; set; }
+        double ElapsedTime { get; set; }
 
         /// <summary>
         /// 性能测试结果
